move operation files to tri_mesh subfolder and rename them according to the class names
