+++ src/wmtk/operations/tri_mesh/EdgeSplit.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <optional>
 #include <wmtk/TriMesh.hpp>
-#include "Operation.hpp"
+#include "../Operation.hpp"
 
 namespace wmtk::operations {
 namespace tri_mesh {
