+++ src/wmtk/operations/tri_mesh/EdgeSplitAtMidpoint.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <wmtk/TriMesh.hpp>
-#include "Operation.hpp"
+#include "../Operation.hpp"
 
 namespace wmtk::operations {
 namespace tri_mesh {
