+++ src/wmtk/operations/tri_mesh/EdgeSwap.hpp
@@ -1,7 +1,7 @@
 
 #pragma once
 #include <optional>
-#include "Operation.hpp"
+#include "../Operation.hpp"
 
 namespace wmtk::operations {
 namespace tri_mesh {
